SDK(Notifications): Update script to help automate deployments
After incorporating the notifications panel code into the Calypso
repository in #26757 the existing deployment process became more
complicated than it had been.

In this patch we're updating the SDK script for the notifications
panel to account for changes to the SDK system and in order to
preserve some existing filenames and norms.

Additionally we're adding the generation of the output bundle into
the CircleCI suite so that it's available whenever we make changes.
This should serve as a platform on which we can add new tests and
copy the code directly to deployment on the iframe version of the app.